+++ src/test/java/net/kaczmarzyk/spring/data/jpa/domain/ConjunctionTest.java
@@ -15,17 +15,18 @@
  */
 package net.kaczmarzyk.spring.data.jpa.domain;
 
-import static net.kaczmarzyk.spring.data.jpa.CustomerBuilder.customer;
-import static org.assertj.core.api.Assertions.*;
-import java.text.ParseException;
-import java.util.List;
-
+import net.kaczmarzyk.spring.data.jpa.Customer;
+import net.kaczmarzyk.spring.data.jpa.IntegrationTestBase;
+import nl.jqno.equalsverifier.EqualsVerifier;
+import nl.jqno.equalsverifier.Warning;
 import org.junit.Before;
 import org.junit.Test;
 
-import net.kaczmarzyk.spring.data.jpa.Customer;
-import net.kaczmarzyk.spring.data.jpa.IntegrationTestBase;
-import net.kaczmarzyk.spring.data.jpa.utils.Converter;
+import java.text.ParseException;
+import java.util.List;
+
+import static net.kaczmarzyk.spring.data.jpa.CustomerBuilder.customer;
+import static org.assertj.core.api.Assertions.assertThat;
 
 
 public class ConjunctionTest extends IntegrationTestBase {
@@ -55,5 +56,14 @@ public class ConjunctionTest extends IntegrationTestBase {
         assertThat(result)
             .hasSize(2)
             .containsOnly(homerSimpson, margeSimpson);
+    }
+
+    @Test
+    public void equalsAndHashCodeContract() {
+        EqualsVerifier.forClass(Conjunction.class)
+                .usingGetClass()
+                .suppress(Warning.NONFINAL_FIELDS)
+                .withIgnoredFields("queriesWithInitializedFakes")
+                .verify();
     }
 }